+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 - **BREAKING CHANGE**: `solarwindsprocessor` - Removed deprecated root-level `extension` field. Use `collector_attributes_decoration.extension` instead.
+- `swok8sworkloadtypeprocessor` Improve detection for workloads from different namespaces
 
 ## v0.136.2
 - No changes